Upgrade Test Framework to 0.0.4
