+++ .github/workflows/ci.yml
@@ -63,10 +63,10 @@ jobs:
       - name: Build with Maven
         run: |
           mvn -V -B -s .github/mvn-settings.xml verify -Dvalidate-format -DskipTests -DskipITs
-  build-jvm:
-    name: JVM build - latest Quarkus
+  linux-build-jvm-latest:
+    name: PR - Linux - JVM build - Latest Version
     runs-on: ubuntu-latest
-    needs: build-dependencies
+    needs: [ build-dependencies, linux-validate-format ]
     strategy:
       matrix:
         java: [ 11 ]
@@ -108,16 +108,16 @@ jobs:
       - name: Zip Artifacts
         if: failure()
         run: |
-          zip -R artifacts-jvm${{ matrix.java }}.zip '*-reports/*'
+          zip -R artifacts-latest-linux-jvm${{ matrix.java }}.zip '*-reports/*'
       - uses: actions/upload-artifact@v1
         if: failure()
         with:
           name: ci-artifacts
-          path: artifacts-jvm${{ matrix.java }}.zip
-  build-native:
-    name: Native build - latest Quarkus
+          path: artifacts-latest-linux-jvm${{ matrix.java }}.zip
+  linux-build-native-latest:
+    name: PR - Linux - Native build - Latest Version
     runs-on: ubuntu-latest
-    needs: build-dependencies
+    needs: [ build-dependencies, linux-validate-format ]
     strategy:
       matrix:
         java: [ 11 ]
@@ -179,9 +179,52 @@ jobs:
       - name: Zip Artifacts
         if: failure()
         run: |
-          zip -R artifacts-native${{ matrix.java }}.zip '*-reports/*'
+          zip -R artifacts-latest-linux-native${{ matrix.java }}.zip '*-reports/*'
       - uses: actions/upload-artifact@v1
         if: failure()
         with:
           name: ci-artifacts
-          path: artifacts-native${{ matrix.java }}.zip
+          path: artifacts-latest-linux-native${{ matrix.java }}.zip
+  windows-build-jvm-latest:
+    name: PR - Windows - JVM build - Latest Version
+    runs-on: windows-latest
+    needs: [ build-dependencies, linux-validate-format ]
+    strategy:
+      matrix:
+        java: [ 11 ]
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/cache@v1
+        with:
+          path: ~/.m2/repository
+          key: ${{ runner.os }}-maven-${{ hashFiles('**/pom.xml') }}
+          restore-keys: |
+            ${{ runner.os }}-maven-
+      - name: Install JDK {{ matrix.java }}
+        uses: joschi/setup-jdk@e87a7cec853d2dd7066adf837fe12bf0f3d45e52
+        with:
+          java-version: ${{ matrix.java }}
+      - name: Download Maven Repo
+        uses: actions/download-artifact@v1
+        with:
+          name: maven-repo
+          path: .
+      - name: Extract Maven Repo
+        shell: bash
+        run: tar -xzf maven-repo.tgz -C ~
+      - name: Build in JVM mode
+        shell: bash
+        run: |
+          mvn -fae -s .github/mvn-settings.xml clean verify
+      - name: Zip Artifacts
+        shell: bash
+        if: failure()
+        run: |
+          # Disambiguate windows find from cygwin find
+          /usr/bin/find . -name '*-reports/*' -type d | tar -czf artifacts-latest-windows-jvm${{ matrix.java }}.tar -T -
+      - name: Archive artifacts
+        if: failure()
+        uses: actions/upload-artifact@v1
+        with:
+          name: ci-artifacts
+          path: artifacts-latest-windows-jvm${{ matrix.java }}.tar
