+++ .github/workflows/daily.yaml
@@ -35,8 +35,8 @@ jobs:
         with:
           name: maven-repo
           path: maven-repo.tgz
-  build-jvm:
-    name: Daily build - JVM
+  linux-build-jvm-latest:
+    name: Daily - Linux - JVM build - Latest Version
     runs-on: ubuntu-latest
     needs: build-dependencies
     strategy:
@@ -87,8 +87,8 @@ jobs:
         with:
           name: ci-artifacts
           path: artifacts-jvm${{ matrix.java }}.zip
-  build-native:
-    name: Daily build - Native
+  linux-build-native-latest:
+    name: Daily - Linux - Native build - Latest Version
     runs-on: ubuntu-latest
     needs: build-dependencies
     strategy:
@@ -143,3 +143,105 @@ jobs:
         with:
           name: ci-artifacts
           path: artifacts-native${{ matrix.java }}.zip
+  windows-build-jvm-latest:
+    name: Daily - Windows - JVM build - Latest Version
+    runs-on: windows-latest
+    needs: build-dependencies
+    strategy:
+      matrix:
+        java: [ 11 ]
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/cache@v1
+        with:
+          path: ~/.m2/repository
+          key: ${{ runner.os }}-maven-${{ hashFiles('**/pom.xml') }}
+          restore-keys: |
+            ${{ runner.os }}-maven-
+      - name: Install JDK {{ matrix.java }}
+        uses: joschi/setup-jdk@e87a7cec853d2dd7066adf837fe12bf0f3d45e52
+        with:
+          java-version: ${{ matrix.java }}
+      - name: Download Maven Repo
+        uses: actions/download-artifact@v1
+        with:
+          name: maven-repo
+          path: .
+      - name: Extract Maven Repo
+        shell: bash
+        run: tar -xzf maven-repo.tgz -C ~
+      - name: Build in JVM mode
+        shell: bash
+        run: |
+          mvn -fae -s .github/mvn-settings.xml clean verify
+      - name: Zip Artifacts
+        shell: bash
+        if: failure()
+        run: |
+          # Disambiguate windows find from cygwin find
+          /usr/bin/find . -name '*-reports/*' -type d | tar -czf artifacts-latest-windows-jvm${{ matrix.java }}.tar -T -
+      - name: Archive artifacts
+        if: failure()
+        uses: actions/upload-artifact@v1
+        with:
+          name: ci-artifacts
+          path: artifacts-latest-windows-jvm${{ matrix.java }}.tar
+  windows-build-native-latest:
+    name: Daily - Windows - Native build - Latest Version
+    runs-on: windows-latest
+    needs: build-dependencies
+    strategy:
+      matrix:
+        java: [ 11 ]
+        graalvm-version: [ "21.0.0.java11"]
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/cache@v1
+        with:
+          path: ~/.m2/repository
+          key: ${{ runner.os }}-maven-${{ hashFiles('**/pom.xml') }}
+          restore-keys: |
+            ${{ runner.os }}-maven-
+      - name: Install JDK {{ matrix.java }}
+        uses: joschi/setup-jdk@e87a7cec853d2dd7066adf837fe12bf0f3d45e52
+        with:
+          java-version: ${{ matrix.java }}
+      - name: Download Maven Repo
+        uses: actions/download-artifact@v1
+        with:
+          name: maven-repo
+          path: .
+      - name: Extract Maven Repo
+        shell: bash
+        run: tar -xzf maven-repo.tgz -C ~
+      - name: Install cl.exe
+        uses: ilammy/msvc-dev-cmd@v1
+      - uses: microsoft/setup-msbuild@v1
+      - name: Setup GraalVM
+        id: setup-graalvm
+        uses: DeLaGuardo/setup-graalvm@master
+        with:
+          graalvm-version: ${{ matrix.graalvm-version }}
+      - name: Install native-image component
+        run: |
+          gu.cmd install native-image
+      - name: Configure Pagefile
+        # Increased the page-file size due to memory-consumption of native-image command
+        # For details see https://github.com/actions/virtual-environments/issues/785
+        uses: al-cheb/configure-pagefile-action@v1.2
+      - name: Build in Native mode
+        shell: bash
+        run: |
+          mvn -fae -s .github/mvn-settings.xml clean verify -Dnative -Dquarkus.native.container-build=false
+      - name: Zip Artifacts
+        shell: bash
+        if: failure()
+        run: |
+          # Disambiguate windows find from cygwin find
+          /usr/bin/find . -name '*-reports/*' -type d | tar -czf artifacts-latest-windows-native${{ matrix.java }}.tar -T -
+      - name: Archive artifacts
+        if: failure()
+        uses: actions/upload-artifact@v1
+        with:
+          name: ci-artifacts
+          path: artifacts-latest-windows-native${{ matrix.java }}.tar
