+++ micrometer/prometheus-kafka/src/test/java/io/quarkus/ts/micrometer/prometheus/kafka/BaseOpenShiftAlertEventsIT.java
@@ -127,6 +127,6 @@ public abstract class BaseOpenShiftAlertEventsIT {
     }
 
     protected static void loadServiceMonitor(Service app) {
-        client.apply(app, Paths.get("target/test-classes/service-monitor.yaml"));
+        client.apply(Paths.get("target/test-classes/service-monitor.yaml"));
     }
 }
