+++ micrometer/prometheus/src/test/java/io/quarkus/ts/micrometer/prometheus/HttpServerMetricsIT.java
@@ -2,12 +2,14 @@ package io.quarkus.ts.micrometer.prometheus;
 
 import static io.restassured.RestAssured.given;
 import static io.restassured.RestAssured.when;
+import static org.awaitility.Awaitility.await;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
@@ -17,6 +19,7 @@ import io.quarkus.test.scenarios.QuarkusScenario;
 @QuarkusScenario
 public class HttpServerMetricsIT {
 
+    private static final int ASSERT_METRICS_TIMEOUT_MINUTES = 1;
     private static final List<String> HTTP_SERVER_REQUESTS_METRICS_SUFFIX = Arrays.asList("count", "sum", "max");
 
     private static final String HTTP_SERVER_REQUESTS_METRICS_FORMAT = "http_server_requests_seconds_%s{method=\"GET\",outcome=\"SUCCESS\",status=\"200\",uri=\"%s\",}";
@@ -53,7 +56,9 @@ public class HttpServerMetricsIT {
 
         for (String metricSuffix : HTTP_SERVER_REQUESTS_METRICS_SUFFIX) {
             String metric = String.format(HTTP_SERVER_REQUESTS_METRICS_FORMAT, metricSuffix, PING_PONG_ENDPOINT);
-            assertTrue(metrics.contains(metric), "Expected metric not found: " + metric + ". Metrics: " + metrics);
+            await().ignoreExceptions().atMost(ASSERT_METRICS_TIMEOUT_MINUTES, TimeUnit.MINUTES)
+                    .untilAsserted(() -> assertTrue(metrics.contains(metric),
+                            "Expected metric not found: " + metric + ". Metrics: " + metrics));
         }
     }
 }
