+++ properties/src/main/resources/configsource.prop
