+++ properties/src/main/resources/resources-config.json
@@ -1,7 +1,7 @@
 {
   "resources": [
     {
-      "pattern": "configsource.properties"
+      "pattern": "configsource.prop"
     }
   ]
 }
