+++ quarkus-cli/src/test/java/io/quarkus/ts/quarkus/cli/QuarkusCliCreateApplicationIT.java
@@ -1,6 +1,7 @@
 package io.quarkus.ts.quarkus.cli;
 
 import static io.quarkus.test.utils.AwaitilityUtils.untilAsserted;
+import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.List;
@@ -9,7 +10,6 @@ import java.util.stream.Stream;
 import javax.inject.Inject;
 
 import org.apache.http.HttpStatus;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
@@ -41,9 +41,8 @@ public class QuarkusCliCreateApplicationIT {
         assertTrue(result.isSuccessful(), "The application didn't build on JVM. Output: " + result.getOutput());
 
         // Start using DEV mode
-        // TODO: Dev mode does not work when running on 999-SNAPSHOT. Fixed in Quarkus Test Framework 0.0.4.
-        // app.start();
-        // app.given().get().then().statusCode(HttpStatus.SC_OK);
+        app.start();
+        app.given().get().then().statusCode(HttpStatus.SC_OK);
     }
 
     @Test
@@ -55,13 +54,11 @@ public class QuarkusCliCreateApplicationIT {
         assertInstalledExtensions(app, RESTEASY_SPRING_WEB_EXTENSION);
 
         // Start using DEV mode
-        // TODO: Dev mode does not work when running on 999-SNAPSHOT. Fixed in Quarkus Test Framework 0.0.4.
-        // app.start();
-        // untilAsserted(() -> app.given().get("/greeting").then().statusCode(HttpStatus.SC_OK).and().body(is("Hello Spring")));
+        app.start();
+        untilAsserted(() -> app.given().get("/greeting").then().statusCode(HttpStatus.SC_OK).and().body(is("Hello Spring")));
     }
 
     @Test
-    @Disabled("Dev mode does not work when running on 999-SNAPSHOT. Fixed in Quarkus Test Framework 0.0.4.")
     public void shouldAddAndRemoveExtensions() {
         // Create application
         QuarkusCliRestService app = cliClient.createApplication("app");
