+++ quarkus-cli/src/test/java/io/quarkus/ts/quarkus/cli/QuarkusCliVersionIT.java
@@ -1,6 +1,7 @@
 package io.quarkus.ts.quarkus.cli;
 
 import static io.quarkus.test.utils.AwaitilityUtils.untilAsserted;
+import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -10,7 +11,6 @@ import java.util.stream.Stream;
 import javax.inject.Inject;
 
 import org.apache.http.HttpStatus;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
@@ -52,9 +52,8 @@ public class QuarkusCliVersionIT {
         assertTrue(result.isSuccessful(), "The application didn't build on JVM. Output: " + result.getOutput());
 
         // Start using DEV mode
-        // TODO: Dev mode does not work when running on 999-SNAPSHOT. Fixed in Quarkus Test Framework 0.0.4.
-        // app.start();
-        // app.given().get().then().statusCode(HttpStatus.SC_OK);
+        app.start();
+        app.given().get().then().statusCode(HttpStatus.SC_OK);
     }
 
     @Test
@@ -66,13 +65,11 @@ public class QuarkusCliVersionIT {
         assertInstalledExtensions(app, RESTEASY_SPRING_WEB_EXTENSION);
 
         // Start using DEV mode
-        // TODO: Dev mode does not work when running on 999-SNAPSHOT. Fixed in Quarkus Test Framework 0.0.4.
-        // app.start();
-        // untilAsserted(() -> app.given().get("/greeting").then().statusCode(HttpStatus.SC_OK).and().body(is("Hello Spring")));
+        app.start();
+        untilAsserted(() -> app.given().get("/greeting").then().statusCode(HttpStatus.SC_OK).and().body(is("Hello Spring")));
     }
 
     @Test
-    @Disabled("Dev mode does not work when running on 999-SNAPSHOT. Fixed in Quarkus Test Framework 0.0.4.")
     public void shouldAddAndRemoveExtensions() {
         // Create application
         QuarkusCliRestService app = cliClient.createApplication("app");
