+++ scheduling/quartz/src/test/resources/mysql.properties
@@ -1,6 +1,3 @@
 # MySql
 quarkus.datasource.username=user
 quarkus.datasource.password=user
-
-# Workaround to locate the migration scripts. It will be fixed in Quarkus Test Framework 0.0.4
-quarkus.flyway.locations=/
