+++ security/https/src/test/java/io/quarkus/ts/security/https/secured/BaseAuthzHttpsSecurityIT.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.security.https.enabled;
+package io.quarkus.ts.security.https.secured;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -8,6 +8,8 @@ import java.security.GeneralSecurityException;
 
 import javax.net.ssl.SSLContext;
 
+import org.apache.http.HttpResponse;
+import org.apache.http.HttpStatus;
 import org.apache.http.client.fluent.Executor;
 import org.apache.http.client.fluent.Request;
 import org.apache.http.conn.ssl.DefaultHostnameVerifier;
@@ -18,24 +20,16 @@ import org.junit.jupiter.api.Test;
 
 import io.quarkus.test.bootstrap.Protocol;
 import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.QuarkusScenario;
-import io.quarkus.test.services.QuarkusApplication;
 import io.quarkus.ts.security.https.utils.Certificates;
 import io.quarkus.ts.security.https.utils.HttpsAssertions;
-import io.quarkus.vertx.http.runtime.HttpConfiguration;
 
-@QuarkusScenario
-public class EnabledHttpsSecurityIT {
+public abstract class BaseAuthzHttpsSecurityIT {
     // not using RestAssured because we want 100% control over certificate & hostname verification
 
     static final char[] CLIENT_PASSWORD = "client-password".toCharArray();
 
-    @QuarkusApplication(ssl = true)
-    static RestService app = new RestService()
-            .withProperty("quarkus.http.insecure-requests", HttpConfiguration.InsecureRequests.ENABLED.name());
-
     @Test
-    public void https() throws IOException, GeneralSecurityException {
+    public void httpsAuthenticatedAndAuthorizedClient() throws IOException, GeneralSecurityException {
         SSLContext sslContext = SSLContexts.custom()
                 .setKeyStoreType(Certificates.PKCS12)
                 .loadKeyMaterial(new File(Certificates.CLIENT_KEYSTORE), CLIENT_PASSWORD, CLIENT_PASSWORD)
@@ -46,10 +40,35 @@ public class EnabledHttpsSecurityIT {
                 .setSSLHostnameVerifier(new DefaultHostnameVerifier())
                 .build()) {
 
-            String response = Executor.newInstance(httpClient)
-                    .execute(Request.Get(url(Protocol.HTTPS)))
-                    .returnContent().asString();
-            assertEquals("Hello, use SSL true", response);
+            Executor executor = Executor.newInstance(httpClient);
+
+            assertEquals("Hello CN=client, HTTPS: true, isUser: true, isGuest: false",
+                    executor.execute(Request.Get(url(Protocol.HTTPS))).returnContent().asString());
+
+            assertEquals("Client certificate: CN=client",
+                    executor.execute(Request.Get(urlWithAuthz())).returnContent().asString());
+        }
+    }
+
+    @Test
+    public void httpsAuthenticatedButUnauthorizedClient() throws IOException, GeneralSecurityException {
+        SSLContext sslContext = SSLContexts.custom()
+                .setKeyStoreType(Certificates.PKCS12)
+                .loadKeyMaterial(new File(Certificates.GUESS_CLIENT_KEYSTORE), CLIENT_PASSWORD, CLIENT_PASSWORD)
+                .loadTrustMaterial(new File(Certificates.CLIENT_TRUSTSTORE), CLIENT_PASSWORD)
+                .build();
+        try (CloseableHttpClient httpClient = HttpClients.custom()
+                .setSSLContext(sslContext)
+                .setSSLHostnameVerifier(new DefaultHostnameVerifier())
+                .build()) {
+
+            Executor executor = Executor.newInstance(httpClient);
+
+            assertEquals("Hello CN=guest-client, HTTPS: true, isUser: false, isGuest: true",
+                    executor.execute(Request.Get(url(Protocol.HTTPS))).returnContent().asString());
+
+            HttpResponse response = executor.execute(Request.Get(urlWithAuthz())).returnResponse();
+            assertEquals(HttpStatus.SC_FORBIDDEN, response.getStatusLine().getStatusCode());
         }
     }
 
@@ -82,8 +101,16 @@ public class EnabledHttpsSecurityIT {
                 .setSSLHostnameVerifier(new DefaultHostnameVerifier())
                 .build()) {
 
-            HttpsAssertions.assertTlsHandshakeError(() -> {
-                Executor.newInstance(httpClient).execute(Request.Get(url(Protocol.HTTPS)));
+            Executor executor = Executor.newInstance(httpClient);
+
+            HttpsAssertions.assertTls13OnlyHandshakeError(() -> {
+                String response = executor.execute(Request.Get(url(Protocol.HTTPS))).returnContent().asString();
+                assertEquals("Hello <anonymous>, HTTPS: true, isUser: false, isGuest: false", response);
+            });
+
+            HttpsAssertions.assertTls13OnlyHandshakeError(() -> {
+                HttpResponse response = executor.execute(Request.Get(urlWithAuthz())).returnResponse();
+                assertEquals(HttpStatus.SC_FORBIDDEN, response.getStatusLine().getStatusCode());
             });
         }
     }
@@ -109,10 +136,16 @@ public class EnabledHttpsSecurityIT {
     @Test
     public void http() throws IOException {
         String response = Request.Get(url(Protocol.HTTP)).execute().returnContent().asString();
-        assertEquals("Hello, use SSL false", response);
+        assertEquals("Hello <anonymous>, HTTPS: false, isUser: false, isGuest: false", response);
     }
 
+    protected abstract RestService getApp();
+
     private String url(Protocol protocol) {
-        return app.getHost(protocol) + ":" + app.getPort(protocol) + "/hello/simple";
+        return getApp().getHost(protocol) + ":" + getApp().getPort(protocol) + "/hello/full";
+    }
+
+    private String urlWithAuthz() {
+        return getApp().getHost(Protocol.HTTPS) + ":" + getApp().getPort(Protocol.HTTPS) + "/secured";
     }
 }
