+++ security/https/src/test/resources/windows.properties
@@ -1,0 +1,9 @@
+quarkus.http.ssl.certificate.key-store-file=META-INF\\resources\\server-keystore.pkcs12
+quarkus.http.ssl.certificate.key-store-password=server-password
+quarkus.http.ssl.certificate.trust-store-file=META-INF\\resources\\server-truststore.pkcs12
+quarkus.http.ssl.certificate.trust-store-password=server-password
+quarkus.http.ssl.client-auth=request
+quarkus.ssl.native=true
+quarkus.http.auth.policy.users-only.roles-allowed=user
+quarkus.http.auth.permission.secured-urls.paths=/secured/*
+quarkus.http.auth.permission.secured-urls.policy=users-only
