+++ sql-db/panache-flyway/src/main/resources/application.properties
@@ -30,6 +30,3 @@ quarkus.flyway.schemas=test
 %agroal_pool_test.quarkus.datasource.jdbc.acquisition-timeout=60S
 %agroal_pool_test.quarkus.datasource.jdbc.validation-query-sql=SELECT CURRENT_TIMESTAMP
 %agroal_pool_test.quarkus.datasource.jdbc.new-connection-sql=SELECT CURRENT_TIMESTAMP
-
-quarkus.hibernate-orm.log.bind-parameters=true
-quarkus.hibernate-orm.log.sql=true
